[FFM-5352]: Stream errors cause the SDK to spam ff server requests
This change introduces a backoff retry on stream failure.
It also signals that the poller should start on any stream failures,
pausing the poller if the stream successfully reconnects.

On a stream error the SDK starts spamming the ff-server with requests
as fast as it can.  This has resulte in 429 being returned to some
customers.