FFM-6932 - SDK seems to disconnect from the relay proxy
What
Add timeout parameter of 60 seconds to HTTP request to force socket to time out if no data arrives.
The timeout will trigger an exception so that the poller is restarted. Also before putting
the poller to sleep when stream is enabled, make sure we retrieve the flags.

Why
Customers are reporting SSE updates stopping and flags getting out of date

Testing
Manual - ensured that after connection was lost to SSE stream that the poller is restarted